Looking at issue #9
Loading the extension into the production VS causes a hang when the solution
is opened.

This does not occur in the experimental instance.

Investigation shows that disabling the default implementation of
ConfigureApplicationContainer in the default Nancy bootstrapper makes the
problem go away.

Since this reflects across the appdomain, and essentially registers everything
in the container, I'm attempting to load the server into a separate appdomain.

It's not working :(

So: I'm going to go back to the state of play where we had a separate process
hosting the web server. I'll use Nancy self-host, rather than IIS Express, though.